Remove `#[must_use]` from most of types
Apparently warnings about unused types is annoying when the type
is meant to be read from. So I removed `#[must_use]` on such types,
but I did add it to methods returning such structs, so that warnings are
actually emitted when they're not even read from. Also, I didn't touch
input types, because the user can only write to them but not read and
they have to be passed into `tbot`.