Fall back `can_pin_messages` to `true`
This field is not sent for channel administrators — in channels, all
administrators can pin messages. To keep compatibility, when the field
is not present, I decided to fall back the field to `true`, because the
field _is_ sent for group administrators. In v0.6, the next breaking
release, this field needs to be wrapped in an `Option`.