Move `Bot`'s inner fields into a separate struct
This commit moves the `token` and `client` fields on `Bot` into a
separate struct, `InnerBot`, which is stored behind an `Arc` in `Bot`.
At the same time, `Token`'s inner `Arc` is removed, as well as the `Arc`
wrapping the client. Overall, this commit reduces the number of used `Arc`
in half.

As a result, the token and the client can be passed to methods as one
argument, `&InnerBot`, reducing code size. This also opens up the
possibility to easily extend the struct, e.g. for implementing
rate-limiting.